feat: page dims only after 1s of loading
